ml-service/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
pydantic==2.10.6
numpy==2.2.2
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
python-multipart==0.0.20
Pillow==11.1.0
